+++ bruges/attribute/__init__.py
@@ -1,11 +1,11 @@
-# -*- coding: utf-8 -*-
 """
 :copyright: 2015 Agile Geoscience
 :license: Apache 2.0
 """
 from .energy import energy
 
-from .similarity import similarity
+from .discontinuity import discontinuity
+from .discontinuity import similarity
 
 from .dipsteer import dipsteer
 
@@ -20,3 +20,4 @@ from .complex import instantaneous_phase
 from .complex import instantaneous_frequency
 from .complex import quadrature
 
+from .horizon import *