+++ bruges/attribute/horizon.py
@@ -1,0 +1,10 @@
+"""
+Various horizon attributes.
+
+:copyright: 2021 Agile Scientific
+:license: Apache 2.0
+"""
+import numpy as np
+
+def dip(arr):
+    raise NotImplementedError